+++ src/Common/src/Abstractions/CasingConventions/EnumExtensions.cs
@@ -1,8 +1,8 @@
-﻿// Licensed to the .NET Foundation under one or more agreements.
+// Licensed to the .NET Foundation under one or more agreements.
 // The .NET Foundation licenses this file to you under the Apache 2.0 License.
 // See the LICENSE file in the project root for more information.
 
-namespace Steeltoe.Common.Util;
+namespace Steeltoe.Common.CasingConventions;
 
 public static class EnumExtensions
 {
